Touch code redone.
Touch code didn't work properly before, with touches getting confused
and weird results happening such as jumping, then holding down move,
and the jump button not being unpressed when you let go of it.

This touch code works properly by identifying each touch by its ID and
assigning an ID to a button. The proof is in the pudding. The buttons
now press properly, even when holding down the Dpad and doing a jump.
The log output also verifies the touch code is now correct.