Tweaking button issues.
There are still some button issues. Touches would linger around even
after a finger was removed, so now I check to see if the touch is in a
phase I would want touching the button. This seems to work, but now
there's the opposite problem of touches not registering at all. GRUMBLE.